Pixel Rendering
Fixed game crash when ending and restarting the application.

Created a render and update method and a timer in the game loop.

Created a pixel array that can be manipulated then set to a bitmap. Once
pixels have been set to the bitmap, the bitmap is drawn to the canvas.