Ball
Added the Ball class to the game that has basic information about the
ball's position and size. Currently the ball just sits on top of the
paddle and follows it around the screen.

Made the Paddle class to move the paddle specific information out of the
BricksView class.

Made a drawBall method in BricksView to draw the ball to the screen.