Touch Events
Finally got some Android devices to develop the touch screen side of the
game. Also fixed a bug that would crash the game when changed to
landscape view. Game still crashes when a keyboard is opened up.

Added walls and made those walls the universal boundary for any size
screen.